Indy's HW1 fixed to take decimal numbers (use float for input instead of int).txt

+++ srijumnon_intouch_indy_hw1_2.2.py
@@ -1,7 +1,8 @@
-T=int(input("T in seconds = "))
+T=float(input("T in seconds = "))
 G=6.67E-11
 M=5.97E24
 R=6371000
 Pi=3.1416
 h=((G*M*(T**2)/(4*(Pi**2)))**(1/3))-R
 print("Altitude in meters =",h)
+
